Fix mifare authentication for multiple blocks
Around line 1320 in rfid_asic_rc632.c the CRYPTO1 bit in the                                     
RC632 is disabled. The RC632 docs state that this bit can only be set                                      
by a successfull AUTHENT2 command, but may be cleared by the user to                                       
disable encryption thereafter (but they don't say why anyone would want                                    
to do that). (Henryk Ploetz)


git-svn-id: https://svn.gnumonks.org/trunk/librfid@1814 e0336214-984f-0b4b-a45f-81c69e1f0ede